单独设置了一个包存储trajectories

+++ src/plan_manager/src/path_searching/topo_prm.cpp
@@ -500,7 +500,7 @@ vector<Eigen::Vector3d> TopoPRM::discretizePath(const vector<Eigen::Vector3d>& p
 
         // find the range cur_l in
         int idx = -1;
-        for (int j = 0; j < len_list.size() - 1; ++j) 
+        for (size_t j = 0; j < len_list.size() - 1; ++j) 
         {
             if (cur_l >= len_list[j] - 1e-4 && cur_l <= len_list[j + 1] + 1e-4) 
             {
@@ -604,16 +604,16 @@ void TopoPRM::shortcutPaths()
     if (parallel_shortcut_) 
     {
         vector<thread> short_threads;
-        for (int i = 0; i < raw_paths_.size(); ++i)
+        for (size_t i = 0; i < raw_paths_.size(); ++i)
         {
             short_threads.push_back(std::thread(&TopoPRM::shortcutPath, this, raw_paths_[i], i, 1));
         }
-        for (int i = 0; i < raw_paths_.size(); ++i) 
+        for (size_t i = 0; i < raw_paths_.size(); ++i) 
         {
             short_threads[i].join();
         }
     } else {
-        for (int i = 0; i < raw_paths_.size(); ++i)
+        for (size_t i = 0; i < raw_paths_.size(); ++i)
         {
             shortcutPath(raw_paths_[i], i);
         }
@@ -650,7 +650,7 @@ vector<Eigen::Vector3d> TopoPRM::discretizePath(vector<Eigen::Vector3d> path)
         return dis_path;
     }
 
-    for (int i = 0; i < path.size() - 1; ++i) 
+    for (size_t i = 0; i < path.size() - 1; ++i) 
     {
         segment = discretizeLine(path[i], path[i + 1]);
 
@@ -672,7 +672,7 @@ vector<vector<Eigen::Vector3d>> TopoPRM::discretizePaths(vector<vector<Eigen::Ve
     vector<vector<Eigen::Vector3d>> dis_paths;
     vector<Eigen::Vector3d> dis_path;
 
-    for (int i = 0; i < path.size(); ++i) {
+    for (size_t i = 0; i < path.size(); ++i) {
         dis_path = discretizePath(path[i]);
 
         if (dis_path.size() > 0)
@@ -697,7 +697,7 @@ vector<vector<Eigen::Vector3d>> TopoPRM::searchPaths()
     int min_node_num = 10000, max_node_num = 1;
     vector<vector<int>> path_list(100);
 
-    for(int i=0;i<raw_paths_.size();i++)
+    for(size_t i=0;i<raw_paths_.size();i++)
     {
         if(int(raw_paths_[i].size()) > max_node_num)
         {
@@ -719,10 +719,10 @@ vector<vector<Eigen::Vector3d>> TopoPRM::searchPaths()
 
     // select paths with less nodes
     vector<vector<Eigen::Vector3d>> filter_raw_paths;
-    for(int i= min_node_num; i<=max_node_num;i++)
+    for(size_t i= min_node_num; i<=max_node_num;i++)
     {
         bool reach_max = false;
-        for(int j=0;j < path_list[i].size();j++)
+        for(size_t j=0;j < path_list[i].size();j++)
         {
             filter_raw_paths.push_back(raw_paths_[path_list[i][j]]);
             if(filter_raw_paths.size() >= max_raw_path2_) // 如果路径过多,超出了限制,就break;
@@ -748,13 +748,13 @@ void TopoPRM::depthFirstSearch(vector<GraphNode::Ptr>& vis)
     GraphNode::Ptr cur_node = vis.back();
     
     // check reach goal
-    for(int i=0; i<cur_node->neighbors_.size();i++)
+    for(size_t i=0; i<cur_node->neighbors_.size();i++)
     {
         if(cur_node->neighbors_[i]->id_ == 1)
         {      
             // add this path to paths set
             vector<Eigen::Vector3d> path;
-            for(int j=0 ; j <vis.size(); j++)
+            for(size_t j=0 ; j <vis.size(); j++)
             {
                 path.push_back(vis[j]->pos_);
             }
@@ -768,14 +768,14 @@ void TopoPRM::depthFirstSearch(vector<GraphNode::Ptr>& vis)
         }
     }
 
-    for(int i=0 ; i < cur_node->neighbors_.size(); i++)
+    for(size_t i=0 ; i < cur_node->neighbors_.size(); i++)
     {
         // skip reach goal
         if(cur_node->neighbors_[i]->id_ == 1) continue;
 
         // skip already visited node
         bool revisit = false;
-        for(int j=0; j <vis.size();j++)
+        for(size_t j=0; j <vis.size();j++)
         {
             if(cur_node->neighbors_[i]->id_ == vis[j]->id_)
             {
